Refactored Google provider into a package containing both Hybrid and OAuth2 authentication methods.
