+++ velruse/providers/google/hybrid.py
@@ -14,9 +14,6 @@ from urlparse import parse_qs
 import oauth2 as oauth
 from openid.extensions import ax
 
-from pyramid.security import NO_PERMISSION_REQUIRED
-
-from velruse.api import register_provider
 from velruse.providers.oid_extensions import OAuthRequest
 from velruse.providers.oid_extensions import UIRequest
 from velruse.providers.openid import (
@@ -33,41 +30,6 @@ GOOGLE_OAUTH = 'https://www.google.com/accounts/OAuthGetAccessToken'
 
 class GoogleAuthenticationComplete(OpenIDAuthenticationComplete):
     """Google auth complete"""
-
-
-def includeme(config):
-    config.add_directive('add_google_login', add_google_login)
-
-
-def add_google_login(config,
-                     attrs=None,
-                     realm=None,
-                     storage=None,
-                     consumer_key=None,
-                     consumer_secret=None,
-                     scope=None,
-                     login_path='/login/google',
-                     callback_path='/login/google/callback',
-                     name='google'):
-    """
-    Add a Google login provider to the application.
-
-    OpenID parameters: attrs, realm, storage
-
-    OAuth parameters: consumer_key, consumer_secret, scope
-    """
-    provider = GoogleConsumer(name, attrs, realm, storage,
-                              consumer_key, consumer_secret, scope)
-
-    config.add_route(provider.login_route, login_path)
-    config.add_view(provider, attr='login', route_name=provider.login_route,
-                    permission=NO_PERMISSION_REQUIRED)
-
-    config.add_route(provider.callback_route, callback_path,
-                     use_global_views=True,
-                     factory=provider.callback)
-
-    register_provider(config, name, provider)
 
 
 class GoogleConsumer(OpenIDConsumer):
