+++ velruse/providers/google/oauth2.py
@@ -4,69 +4,23 @@ from json import loads
 import requests
 
 from pyramid.httpexceptions import HTTPFound
-from pyramid.security import NO_PERMISSION_REQUIRED
 
 from velruse.api import (
     AuthenticationComplete,
     AuthenticationDenied,
-    register_provider,
 )
 from velruse.exceptions import ThirdPartyFailure
-from velruse.settings import ProviderSettings
 from velruse.utils import flat_url
+
+
+GOOGLE_OAUTH2_DOMAIN = 'accounts.google.com'
 
 
 class Google2AuthenticationComplete(AuthenticationComplete):
     """Google OAuth 2.0 auth complete"""
 
 
-def includeme(config):
-    config.add_directive('add_google2_login', add_google2_login)
-    config.add_directive('add_google2_login_from_settings',
-                         add_google2_login_from_settings)
-
-
-def add_google2_login_from_settings(config, prefix='velruse.google2.'):
-    settings = config.registry.settings
-    p = ProviderSettings(settings, prefix)
-    p.update('consumer_key', required=True)
-    p.update('consumer_secret', required=True)
-    p.update('scope')
-    p.update('login_path')
-    p.update('callback_path')
-    config.add_google2_login(**p.kwargs)
-
-
-def add_google2_login(config,
-                     consumer_key,
-                     consumer_secret,
-                     scope=None,
-                     login_path='/login/google2',
-                     callback_path='/login/google2/callback',
-                     secure=True,
-                     domain='accounts.google.com',
-                     name='google2'):
-    """
-    Add a Google OAuth 2.0 login provider to the application.
-    """
-    provider = Google2Provider(name,
-                               consumer_key,
-                               consumer_secret,
-                               scope,
-                               domain)
-
-    config.add_route(provider.login_route, login_path)
-    config.add_view(provider, attr='login', route_name=provider.login_route,
-                    permission=NO_PERMISSION_REQUIRED)
-
-    config.add_route(provider.callback_route, callback_path,
-                     use_global_views=True,
-                     factory=provider.callback)
-
-    register_provider(config, name, provider)
-
-
-class Google2Provider(object):
+class GoogleOAuth2Provider(object):
 
     profile_scope = 'https://www.googleapis.com/auth/userinfo.profile'
     email_scope = 'https://www.googleapis.com/auth/userinfo.email'
@@ -75,14 +29,13 @@ class Google2Provider(object):
                  name,
                  consumer_key,
                  consumer_secret,
-                 scope,
-                 domain):
+                 scope):
         self.name = name
-        self.type = 'google2'
+        self.type = 'google_oauth2'
         self.consumer_key = consumer_key
         self.consumer_secret = consumer_secret
         self.protocol = 'https'
-        self.domain = domain
+        self.domain = GOOGLE_OAUTH2_DOMAIN
 
         self.login_route = 'velruse.%s-login' % name
         self.callback_route = 'velruse.%s-callback' % name
@@ -96,9 +49,6 @@ class Google2Provider(object):
         scope = ' '.join(request.POST.getall('scope')) or self.scope
 
         approval_prompt = request.POST.get('approval_prompt', 'auto')
-
-        if request.POST.get('end_point'):
-            request.session['endpoint'] = request.POST['end_point']
 
         auth_url = flat_url(
             '%s://%s/o/oauth2/auth' % (self.protocol, self.domain),
